Add script to add doc comments
Script applies a patch that adds Rust doc comments to a file, accounting
for types that may have changed. For example, the script will detect if
a Rust enum has been changed to a series of constant. The script
includes unit tests.

Also add documentation on how to update doc comments.