Support x86 detail
Replace Representative trait and macros with impl_PartialEq_repr_fields
macro. Instead of instantiating a tuple and comparing the tuple to impl
PartialEq, we now impl PartialEq directly by comparing all fields. This
has the nice side-effect of only needing one macro call and not needing
to pass the type of the fields.